Fix handling of plural strings
This commit makes quite an overhaul. It removes the usage of
node-gettext because it provides overhead on top of gettext-parser that
we don't need. Instead we use gettext-parser directly and iterate over
the data structure it produces. This ensures that plural translations
are also properly merged.

To ensure plurality is handled correctly we implement verification of
the Plural-Forms header. To deal with the fact that brackets and spaces
can cause Plural-Forms strings to look very different even if they mean
the same we just a simple javascript expression parser library. This
turns the plural form expression into a simple operator tree that
ignores unneccessary nesting and can easily be compared.